Move some sped/piggyback things to separate files
